CHANGE: Use int instead of uint16 for local loop variables when iterating through collections (16-bit integers are converted to 32-bit integers before being computed, so it is better to use "int" in stack-allocated variables not sored on the heap at any stage).

+++ tpl/fixed_list_tpl.h
@@ -16,11 +16,11 @@
 #define TPL_FIXED_LIST_H
 
 #ifndef ITERATE
-#define ITERATE(collection,enumerator) for(uint16 enumerator = 0; enumerator < (collection).get_count(); enumerator++)
+#define ITERATE(collection,enumerator) for(int enumerator = 0; enumerator < (collection).get_count(); enumerator++)
 #endif
 
 #ifndef ITERATE_PTR
-#define ITERATE_PTR(collection,enumerator) for(uint16 enumerator = 0; enumerator < (collection)->get_count(); enumerator++)
+#define ITERATE_PTR(collection,enumerator) for(int enumerator = 0; enumerator < (collection)->get_count(); enumerator++)
 #endif 
 
 #include <typeinfo>
